Fix bug in `remove_from_list`
Free block lists in emalloc come in two forms: singly linked
and doubly-linked. `remove_from_list` failed to deal with
singly linked case when the block to be removed wat not at
the list head.

With the fix in `possibly_merge`, this bug can be easily
reproduced with multi-threaded test cases.

Signed-off-by: xxu36 <xiaofeng.xu@intel.com>